+++ backend/app/configuration/domains/schema_response.py
@@ -1,8 +1,9 @@
 from typing import Sequence
 from uuid import UUID
+from warnings import deprecated
 
 from app.data_products.schema import DataProduct
-from app.datasets.schema import Dataset
+from app.datasets.schema import Dataset, OutputPort
 from app.shared.schema import ORMModel
 
 
@@ -12,18 +13,50 @@ class BaseDomainGet(ORMModel):
     description: str
 
 
-class DomainGet(BaseDomainGet):
+@deprecated("Use DomainGetResponse instead")
+class DomainGetOld(BaseDomainGet):
     data_products: list[DataProduct]
     datasets: list[Dataset]
 
 
-class DomainsGetItem(BaseDomainGet):
+class GetDomainResponse(BaseDomainGet):
+    data_products: list[DataProduct]
+    output_ports: list[OutputPort]
+
+    @classmethod
+    def from_domain_get_old(cls, d: DomainGetOld) -> "GetDomainResponse":
+        return GetDomainResponse(
+            id=d.id,
+            name=d.name,
+            description=d.description,
+            data_products=d.data_products,
+            output_ports=d.datasets,
+        )
+
+
+@deprecated("Use GetDomainsItem instead")
+class GetDomainsItemOld(BaseDomainGet):
     data_product_count: int
     dataset_count: int
 
 
-class DomainsGet(ORMModel):
-    domains: Sequence[DomainsGetItem]
+class GetDomainsItem(BaseDomainGet):
+    data_product_count: int
+    output_port_count: int
+
+    @classmethod
+    def from_get_domains_item_old(cls, d: GetDomainsItemOld) -> "GetDomainsItem":
+        return GetDomainsItem(
+            id=d.id,
+            name=d.name,
+            description=d.description,
+            data_product_count=d.data_product_count,
+            output_port_count=d.dataset_count,
+        )
+
+
+class GetDomainsResponse(ORMModel):
+    domains: Sequence[GetDomainsItem]
 
 
 class CreateDomainResponse(ORMModel):