+++ backend/app/data_outputs_datasets/router.py
@@ -4,6 +4,7 @@ from uuid import UUID
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 
+from app.authorization.role_assignments.enums import DecisionStatus
 from app.core.auth.auth import get_authenticated_user
 from app.core.authz import Action, Authorization, DataOutputDatasetAssociationResolver
 from app.core.aws.refresh_infrastructure_lambda import RefreshInfrastructureLambda
@@ -14,7 +15,6 @@ from app.events.schema import CreateEvent
 from app.events.service import EventService
 from app.notifications.service import NotificationService
 from app.pending_actions.schema import DataOutputDatasetPendingAction
-from app.role_assignments.enums import DecisionStatus
 from app.users.schema import User
 
 router = APIRouter(