+++ backend/app/data_products/model.py
@@ -5,14 +5,16 @@ from sqlalchemy import Column, Enum, ForeignKey, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, Session, mapped_column, relationship
 
+from app.authorization.role_assignments.data_product.model import (
+    DataProductRoleAssignment,
+)
+from app.authorization.role_assignments.enums import DecisionStatus
 from app.configuration.data_product_settings.model import DataProductSettingValue
 from app.configuration.data_product_types.model import DataProductType
 from app.configuration.tags.model import Tag, tag_data_product_table
 from app.data_products.status import DataProductStatus
 from app.data_products_datasets.model import DataProductDatasetAssociation
 from app.database.database import Base, ensure_exists
-from app.role_assignments.data_product.model import DataProductRoleAssignment
-from app.role_assignments.enums import DecisionStatus
 from app.shared.model import BaseORM
 
 if TYPE_CHECKING: