+++ backend/app/datasets/schema.py
@@ -1,15 +1,21 @@
 from uuid import UUID
+from warnings import deprecated
 
-from app.datasets.enums import DatasetAccessType
-from app.datasets.status import DatasetStatus
+from app.datasets.enums import OutputPortAccessType
+from app.datasets.status import OutputPortStatus
 from app.shared.schema import ORMModel
 
 
-class Dataset(ORMModel):
+class OutputPort(ORMModel):
     id: UUID
     name: str
     namespace: str
     description: str
-    status: DatasetStatus
-    access_type: DatasetAccessType
+    status: OutputPortStatus
+    access_type: OutputPortAccessType
     data_product_id: UUID
+
+
+@deprecated("use OutputPort instead")
+class Dataset(OutputPort):
+    pass