+++ backend/app/graph/service.py
@@ -3,6 +3,7 @@ from typing import List
 from sqlalchemy import select
 from sqlalchemy.orm import Session, selectinload
 
+from app.authorization.role_assignments.enums import DecisionStatus
 from app.configuration.domains.model import Domain
 from app.core.logging import logger
 from app.data_products.model import DataProduct
@@ -10,7 +11,6 @@ from app.datasets.model import Dataset
 from app.graph.edge import Edge
 from app.graph.graph import Graph
 from app.graph.node import Node, NodeData, NodeType
-from app.role_assignments.enums import DecisionStatus
 
 
 class GraphService: