+++ backend/tests/app/users/test_router.py
@@ -1,7 +1,7 @@
 import pytest
 
+from app.authorization.roles.schema import Scope
 from app.core.authz.actions import AuthorizationAction
-from app.roles.schema import Scope
 from app.settings import settings
 from tests.factories import UserFactory
 from tests.factories.role import RoleFactory