+++ backend/tests/factories/role.py
@@ -3,11 +3,11 @@ from typing import Final
 import factory
 from faker import Faker
 
+from app.authorization.roles import ADMIN_UUID
+from app.authorization.roles.model import Role
+from app.authorization.roles.schema import Prototype, Scope
 from app.core.authz.actions import AuthorizationAction
 from app.core.authz.authorization import Authorization
-from app.roles import ADMIN_UUID
-from app.roles.model import Role
-from app.roles.schema import Prototype, Scope
 from tests import test_session
 
 faker: Final[Faker] = Faker()